add commit count
As the Github api can only return a maximum of 100 items per API call,
this function instead uses ``git rev-list`` to find the number of commits
in the repository that the bot resides.

The function also uses the first hit on Google for "cool json api", so
the commit count may be littered with [cool number facts](http://numbersapi.com).