Add tag functionality to ~issue
The bot can now return the tags on associated issues.

Each tag on Github can be assigned a color. Due to the difficulty
converting hexadecimal color codes to that of mIRC, the bot will
not use the same colors as on Github: but the colors will remain
consistent (so all 'profile' issues will have 'profile' colored in
the same fashion).